elearning/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[argon2]
python-dotenv
apscheduler
pydantic[email]
pydantic-settings
pydantic==2.8.2
pydantic-settings==2.3.4
pymongo==4.8.0
redis==5.0.4
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pytest==8.2.0
httpx==0.27.0
python-multipart

